updating powershell for swagger and fixing misc bugs

+++ src/ResourceManager/Dns/Commands.Dns/Zones/RemoveAzureDnsZone.cs
@@ -21,6 +21,9 @@ using ProjectResources = Microsoft.Azure.Commands.Dns.Properties.Resources;
 
 namespace Microsoft.Azure.Commands.Dns
 {
+    using System;
+    using Rest.Azure;
+
     /// <summary>
     /// Deletes an existing zone.
     /// </summary>
@@ -50,53 +53,74 @@ namespace Microsoft.Azure.Commands.Dns
 
         public override void ExecuteCmdlet()
         {
-            bool deleted = false;
-            DnsZone zoneToDelete = null;
+            bool deleted = true;
+            bool overwrite = this.Overwrite.IsPresent || this.ParameterSetName != "Object";
 
-            if (this.ParameterSetName == "Fields")
+            // There is a bug in sdk where it doesn't handle non existing zones on delete. Hence, handling the condition in powershell
+            var zoneToDelete = (this.ParameterSetName != "Object")
+                ? GetDnsZoneHandleNonExistentZone(this.Name, this.ResourceGroupName)
+                : this.Zone;
+
+            if (zoneToDelete != null)
             {
-                zoneToDelete = new DnsZone 
-                {
-                    Name = this.Name,
-                    ResourceGroupName = this.ResourceGroupName,
-                    Etag = null,
-                };
-            }
-            else if (this.ParameterSetName == "Object")
-            {
-                if ((string.IsNullOrWhiteSpace(this.Zone.Etag) || this.Zone.Etag == "*") && !this.Overwrite.IsPresent)
+                if ((string.IsNullOrWhiteSpace(zoneToDelete.Etag) || zoneToDelete.Etag == "*") && !overwrite)
                 {
                     throw new PSArgumentException(string.Format(ProjectResources.Error_EtagNotSpecified, typeof(DnsZone).Name));
                 }
 
-                zoneToDelete = this.Zone;
+                if (zoneToDelete.Name != null && zoneToDelete.Name.EndsWith("."))
+                {
+                    zoneToDelete.Name = zoneToDelete.Name.TrimEnd('.');
+                    this.WriteWarning(string.Format("Modifying zone name to remove terminating '.'.  Zone name used is \"{0}\".", zoneToDelete.Name));
+                }
+
+                ConfirmAction(
+                    Force.IsPresent,
+                    string.Format(ProjectResources.Confirm_RemoveZone, zoneToDelete.Name),
+                    ProjectResources.Progress_RemovingZone,
+                    this.Name,
+                    () => { deleted = DnsClient.DeleteDnsZone(zoneToDelete, overwrite); });
+
+                if (deleted)
+                {
+                    WriteVerbose(ProjectResources.Success);
+                    WriteVerbose(string.Format(ProjectResources.Success_RemoveZone, zoneToDelete.Name, zoneToDelete.ResourceGroupName));
+                }
             }
-
-            if (zoneToDelete.Name != null && zoneToDelete.Name.EndsWith("."))
-            {
-                zoneToDelete.Name = zoneToDelete.Name.TrimEnd('.');
-                this.WriteWarning(string.Format("Modifying zone name to remove terminating '.'.  Zone name used is \"{0}\".", zoneToDelete.Name));
-            }
-
-            bool overwrite = this.Overwrite.IsPresent || this.ParameterSetName != "Object";
-
-            ConfirmAction(
-                Force.IsPresent,
-                string.Format(ProjectResources.Confirm_RemoveZone, zoneToDelete.Name),
-                ProjectResources.Progress_RemovingZone,
-                this.Name,
-                () => { deleted = DnsClient.DeleteDnsZone(zoneToDelete, overwrite); });
-
-            if (deleted)
+            else
             {
                 WriteVerbose(ProjectResources.Success);
-                WriteVerbose(string.Format(ProjectResources.Success_RemoveZone, zoneToDelete.Name, zoneToDelete.ResourceGroupName));
+                WriteWarning(string.Format(ProjectResources.Success_NonExistentZone, this.Name, this.ResourceGroupName));
             }
 
             if (this.PassThru)
             {
                 WriteObject(deleted);
             }
+        }
+
+        public DnsZone GetDnsZoneHandleNonExistentZone(string zoneName, string resourceGroupName)
+        {
+            DnsZone retrievedZone = null;
+            try
+            {
+                if (zoneName.EndsWith("."))
+                {
+                    zoneName = zoneName.TrimEnd('.');
+                    this.WriteWarning(string.Format("Modifying zone name to remove terminating '.'.  Zone name used is \"{0}\".", zoneName));
+                }
+
+                retrievedZone = this.DnsClient.GetDnsZone(zoneName, resourceGroupName);
+            }
+            catch (CloudException exception)
+            {
+                if (exception.Body.Code != "ResourceNotFound")
+                {
+                    throw;    
+                }
+            } 
+
+            return retrievedZone;
         }
     }
 }
